+++ crates/swc_typescript/src/fast_dts/mod.rs
@@ -1,7 +1,7 @@
-use std::{mem::take, sync::Arc};
+use std::{borrow::Cow, mem::take, sync::Arc};
 
 use swc_atoms::Atom;
-use swc_common::{util::take::Take, FileName, Span, Spanned, DUMMY_SP};
+use swc_common::{FileName, Span, Spanned, DUMMY_SP};
 use swc_ecma_ast::{
     AssignPat, BindingIdent, ClassMember, ComputedPropName, Decl, DefaultDecl, ExportDecl,
     ExportDefaultExpr, Expr, Ident, Lit, MethodKind, ModuleDecl, ModuleItem, OptChainBase, Param,
@@ -25,12 +25,21 @@ use swc_ecma_ast::{
     TsKeywordTypeKind, TsLit, TsLitType, TsNamespaceBody, TsParamPropParam, TsPropertySignature,
     TsTupleElement, TsTupleType, TsType, TsTypeAnn, TsTypeElement, TsTypeLit, TsTypeOperator,
     TsTypeOperatorOp, TsTypeRef, VarDecl, VarDeclKind, VarDeclarator,
+    BindingIdent, ComputedPropName, ExportDefaultExpr, Expr, Ident, Lit, ModuleDecl, ModuleItem,
+    OptChainBase, Pat, Program, PropName, TsEntityName, TsKeywordType, TsKeywordTypeKind, TsLit,
+    TsLitType, TsTupleElement, TsType, TsTypeAnn, TsTypeOperator, TsTypeOperatorOp, TsTypeRef,
+    VarDecl, VarDeclKind, VarDeclarator,
 };
 use swc_ecma_visit::{VisitMutWith, VisitWith};
 use type_usage::{TypeRemover, TypeUsageAnalyzer};
 
 use crate::diagnostic::{DtsIssue, SourceRange};
 
+mod class;
+mod decl;
+mod r#enum;
+mod function;
+mod inferrer;
 mod type_usage;
 
 /// TypeScript Isolated Declaration support.
@@ -44,9 +53,9 @@ mod type_usage;
 /// The original code is MIT licensed.
 pub struct FastDts {
     filename: Arc<FileName>,
-    is_top_level: bool,
-    id_counter: u32,
     diagnostics: Vec<DtsIssue>,
+    id_counter: u32,
+    // TODO: strip_internal: bool,
 }
 
 /// Diagnostics
@@ -54,73 +63,46 @@ impl FastDts {
     pub fn new(filename: Arc<FileName>) -> Self {
         Self {
             filename,
-            is_top_level: false,
-            id_counter: 0,
             diagnostics: Vec::new(),
+            id_counter: 0,
         }
     }
 
-    fn mark_diagnostic(&mut self, diagnostic: DtsIssue) {
-        self.diagnostics.push(diagnostic)
-    }
-
-    fn source_range_to_range(&self, range: Span) -> SourceRange {
-        SourceRange {
-            filename: self.filename.clone(),
-            span: range,
-        }
-    }
-
-    fn mark_diagnostic_unable_to_infer(&mut self, range: Span) {
-        self.mark_diagnostic(DtsIssue::UnableToInferType {
-            range: self.source_range_to_range(range),
-        })
-    }
-
-    fn mark_diagnostic_any_fallback(&mut self, range: Span) {
-        self.mark_diagnostic(DtsIssue::UnableToInferTypeFallbackAny {
-            range: self.source_range_to_range(range),
-        })
-    }
-
-    fn mark_diagnostic_unsupported_prop(&mut self, range: Span) {
-        self.mark_diagnostic(DtsIssue::UnableToInferTypeFromProp {
-            range: self.source_range_to_range(range),
+    pub fn mark_diagnostic<T: Into<Cow<'static, str>>>(&mut self, message: T, range: Span) {
+        self.diagnostics.push(DtsIssue {
+            message: message.into(),
+            range: SourceRange {
+                filename: self.filename.clone(),
+                span: range,
+            },
         })
     }
 }
 
 impl FastDts {
     pub fn transform(&mut self, program: &mut Program) -> Vec<DtsIssue> {
-        self.is_top_level = true;
-
+        // 1. Transform. We only keep decls.
         match program {
             Program::Module(module) => {
+                module
+                    .body
+                    .retain(|item| item.as_stmt().map(|stmt| stmt.is_decl()).unwrap_or(true));
+                Self::remove_module_function_overloads(module);
                 self.transform_module_items(&mut module.body);
             }
             Program::Script(script) => {
-                let mut last_function_name: Option<Atom> = None;
-                script.body.retain_mut(|stmt| {
-                    if let Some(fn_decl) = stmt.as_decl().and_then(|decl| decl.as_fn_decl()) {
-                        if fn_decl.function.body.is_some() {
-                            if last_function_name
-                                .as_ref()
-                                .is_some_and(|last_name| last_name == &fn_decl.ident.sym)
-                            {
-                                return false;
-                            }
-                        } else {
-                            last_function_name = Some(fn_decl.ident.sym.clone());
-                        }
-                    }
-                    self.transform_module_stmt(stmt)
-                })
+                script.body.retain(|stmt| stmt.is_decl());
+                Self::remove_script_function_overloads(script);
+                for stmt in script.body.iter_mut() {
+                    self.transform_decl(stmt.as_mut_decl().unwrap());
+                }
             }
             Program::Script(script) => script
                 .body
                 .retain_mut(|stmt| self.transform_module_stmt(stmt)),
         }
 
+        // 2. Remove unused imports and decls
         let mut type_usage_analyzer = TypeUsageAnalyzer::default();
         program.visit_with(&mut type_usage_analyzer);
         program.visit_mut_with(&mut TypeRemover::new(
@@ -135,12 +117,8 @@ impl FastDts {
         let orig_items = take(items);
         let mut new_items = Vec::with_capacity(orig_items.len());
 
-        let mut last_function_name: Option<Atom> = None;
-        let mut is_export_default_function_overloads = false;
-
         for mut item in orig_items {
             match &mut item {
-                // Keep all these
                 ModuleItem::ModuleDecl(
                     ModuleDecl::Import(..)
                     | ModuleDecl::TsImportEquals(_)
@@ -149,439 +127,59 @@ impl FastDts {
                     | ModuleDecl::ExportNamed(_)
                     | ModuleDecl::ExportAll(_),
                 ) => new_items.push(item),
-
                 ModuleItem::Stmt(stmt) => {
-                    if let Some(fn_decl) = stmt.as_decl().and_then(|decl| decl.as_fn_decl()) {
-                        if fn_decl.function.body.is_some() {
-                            if last_function_name
-                                .as_ref()
-                                .is_some_and(|last_name| last_name == &fn_decl.ident.sym)
-                            {
-                                continue;
-                            }
-                        } else {
-                            last_function_name = Some(fn_decl.ident.sym.clone());
-                        }
-                    }
-
-                    if self.transform_module_stmt(stmt) {
-                        new_items.push(item);
-                    }
-                }
-
-                ModuleItem::ModuleDecl(ModuleDecl::ExportDecl(ExportDecl {
-                    span, decl, ..
-                })) => {
-                    if let Some(fn_decl) = decl.as_fn_decl() {
-                        if fn_decl.function.body.is_some() {
-                            if last_function_name
-                                .as_ref()
-                                .is_some_and(|last_name| last_name == &fn_decl.ident.sym)
-                            {
-                                continue;
-                            }
-                        } else {
-                            last_function_name = Some(fn_decl.ident.sym.clone());
-                        }
-                    }
-
-                    if let Some(()) = self.decl_to_type_decl(decl) {
-                        new_items.push(
-                            ExportDecl {
-                                decl: decl.take(),
-                                span: *span,
-                            }
-                            .into(),
-                        );
-                    } else {
-                        self.mark_diagnostic(DtsIssue::UnableToInferType {
-                            range: self.source_range_to_range(*span),
-                        })
-                    }
-                }
-
-                ModuleItem::ModuleDecl(ModuleDecl::ExportDefaultDecl(export)) => {
-                    if let Some(fn_expr) = export.decl.as_fn_expr() {
-                        if is_export_default_function_overloads && fn_expr.function.body.is_some() {
-                            is_export_default_function_overloads = false;
-                            continue;
-                        } else {
-                            is_export_default_function_overloads = true;
-                        }
-                    } else {
-                        is_export_default_function_overloads = false;
-                    }
-
-                    match &mut export.decl {
-                        DefaultDecl::Class(class_expr) => {
-                            self.class_body_to_type(&mut class_expr.class.body);
-                        }
-                        DefaultDecl::Fn(fn_expr) => {
-                            fn_expr.function.body = None;
-                        }
-                        DefaultDecl::TsInterfaceDecl(_) => {}
-                    };
-
+                    self.transform_decl(stmt.as_mut_decl().unwrap());
                     new_items.push(item);
                 }
-
+                ModuleItem::ModuleDecl(ModuleDecl::ExportDecl(expor_decl)) => {
+                    self.transform_decl(&mut expor_decl.decl);
+                    new_items.push(item);
+                }
+                ModuleItem::ModuleDecl(ModuleDecl::ExportDefaultDecl(export)) => {
+                    self.transform_default_decl(&mut export.decl);
+                    new_items.push(item);
+                }
                 ModuleItem::ModuleDecl(ModuleDecl::ExportDefaultExpr(export)) => {
-                    let name = self.gen_unique_name();
-                    let name_ident = Ident::new_no_ctxt(name, DUMMY_SP);
+                    let name_ident = Ident::new_no_ctxt(self.gen_unique_name("_default"), DUMMY_SP);
                     let type_ann = self
-                        .expr_to_ts_type(export.expr.clone(), false, true)
+                        .infer_type_from_expr(&export.expr, false, true)
                         .map(type_ann);
 
-                    if let Some(type_ann) = type_ann {
-                        new_items.push(
-                            VarDecl {
-                                span: DUMMY_SP,
-                                kind: VarDeclKind::Const,
-                                declare: true,
-                                decls: vec![VarDeclarator {
-                                    span: DUMMY_SP,
-                                    name: Pat::Ident(BindingIdent {
-                                        id: name_ident.clone(),
-
-                                        type_ann: Some(type_ann),
-                                    }),
-                                    init: None,
-                                    definite: false,
-                                }],
-                                ..Default::default()
-                            }
-                            .into(),
-                        );
-
-                        new_items.push(
-                            ExportDefaultExpr {
-                                span: export.span,
-                                expr: name_ident.into(),
-                            }
-                            .into(),
-                        )
-                    } else {
-                        new_items.push(
-                            ExportDefaultExpr {
-                                span: export.span,
-                                expr: export.expr.take(),
-                            }
-                            .into(),
-                        )
+                    if type_ann.is_none() {
+                        self.default_export_inferred(export.expr.span());
                     }
+
+                    new_items.push(
+                        VarDecl {
+                            span: DUMMY_SP,
+                            kind: VarDeclKind::Const,
+                            declare: true,
+                            decls: vec![VarDeclarator {
+                                span: DUMMY_SP,
+                                name: Pat::Ident(BindingIdent {
+                                    id: name_ident.clone(),
+                                    type_ann,
+                                }),
+                                init: None,
+                                definite: false,
+                            }],
+                            ..Default::default()
+                        }
+                        .into(),
+                    );
+
+                    new_items.push(
+                        ExportDefaultExpr {
+                            span: export.span,
+                            expr: name_ident.into(),
+                        }
+                        .into(),
+                    )
                 }
             }
         }
 
         *items = new_items;
-    }
-
-    fn transform_module_stmt(&mut self, stmt: &mut Stmt) -> bool {
-        let Stmt::Decl(ref mut decl) = stmt else {
-            return false;
-        };
-
-        match decl {
-            Decl::TsEnum(_) | Decl::Class(_) | Decl::Fn(_) | Decl::Var(_) | Decl::TsModule(_) => {
-                if let Some(()) = self.decl_to_type_decl(decl) {
-                    true
-                } else {
-                    self.mark_diagnostic_unable_to_infer(decl.span());
-                    false
-                }
-            }
-
-            Decl::TsInterface(_) | Decl::TsTypeAlias(_) | Decl::Using(_) => true,
-        }
-    }
-
-    fn expr_to_ts_type(
-        &mut self,
-        e: Box<Expr>,
-        as_const: bool,
-        as_readonly: bool,
-    ) -> Option<Box<TsType>> {
-        match *e {
-            Expr::Array(arr) => {
-                let mut elem_types: Vec<TsTupleElement> = Vec::new();
-
-                for elems in arr.elems {
-                    if let Some(expr_or_spread) = elems {
-                        let span = expr_or_spread.span();
-                        if let Some(ts_expr) =
-                            self.expr_to_ts_type(expr_or_spread.expr, as_const, as_readonly)
-                        {
-                            elem_types.push(ts_tuple_element(ts_expr));
-                        } else {
-                            self.mark_diagnostic_unable_to_infer(span);
-                        }
-                    } else {
-                        // TypeScript converts holey arrays to any
-                        // Example: const a = [,,] -> const a = [any, any, any]
-                        elem_types.push(ts_tuple_element(Box::new(TsType::TsKeywordType(
-                            TsKeywordType {
-                                kind: TsKeywordTypeKind::TsAnyKeyword,
-                                span: DUMMY_SP,
-                            },
-                        ))))
-                    }
-                }
-
-                let mut result = Box::new(TsType::TsTupleType(TsTupleType {
-                    span: arr.span,
-                    elem_types,
-                }));
-
-                if as_readonly {
-                    result = ts_readonly(result);
-                }
-                Some(result)
-            }
-
-            Expr::Object(obj) => {
-                let mut members: Vec<TsTypeElement> = Vec::new();
-
-                // TODO: Prescan all object properties to know which ones
-                // have a getter or a setter. This allows us to apply
-                // TypeScript's `readonly` keyword accordingly.
-
-                for item in obj.props {
-                    match item {
-                        PropOrSpread::Prop(prop_box) => {
-                            let prop = *prop_box;
-                            match prop {
-                                Prop::KeyValue(key_value) => {
-                                    let (key, computed) = match key_value.key {
-                                        PropName::Ident(ident) => {
-                                            (Expr::Ident(ident.into()), false)
-                                        }
-                                        PropName::Str(str_prop) => {
-                                            (Lit::Str(str_prop).into(), false)
-                                        }
-                                        PropName::Num(num) => (Lit::Num(num).into(), true),
-                                        PropName::Computed(computed) => (*computed.expr, true),
-                                        PropName::BigInt(big_int) => {
-                                            (Lit::BigInt(big_int).into(), true)
-                                        }
-                                    };
-
-                                    let init_type = self
-                                        .expr_to_ts_type(key_value.value, as_const, as_readonly)
-                                        .map(type_ann);
-
-                                    members.push(TsTypeElement::TsPropertySignature(
-                                        TsPropertySignature {
-                                            span: DUMMY_SP,
-                                            readonly: as_readonly,
-                                            key: Box::new(key),
-                                            computed,
-                                            optional: false,
-                                            type_ann: init_type,
-                                        },
-                                    ));
-                                }
-                                Prop::Shorthand(_)
-                                | Prop::Assign(_)
-                                | Prop::Getter(_)
-                                | Prop::Setter(_)
-                                | Prop::Method(_) => {
-                                    self.mark_diagnostic_unsupported_prop(prop.span());
-                                }
-                            }
-                        }
-                        PropOrSpread::Spread(_) => {
-                            self.mark_diagnostic(DtsIssue::UnableToInferTypeFromSpread {
-                                range: self.source_range_to_range(item.span()),
-                            })
-                        }
-                    }
-                }
-
-                Some(Box::new(TsType::TsTypeLit(TsTypeLit {
-                    span: obj.span,
-                    members,
-                })))
-            }
-            Expr::Lit(lit) => {
-                if as_const {
-                    maybe_lit_to_ts_type_const(&lit)
-                } else {
-                    maybe_lit_to_ts_type(&lit)
-                }
-            }
-            Expr::TsConstAssertion(ts_const) => self.expr_to_ts_type(ts_const.expr, true, true),
-            Expr::TsSatisfies(satisifies) => {
-                self.expr_to_ts_type(satisifies.expr, as_const, as_readonly)
-            }
-            Expr::TsAs(ts_as) => Some(ts_as.type_ann),
-            Expr::Fn(fn_expr) => {
-                let return_type = fn_expr
-                    .function
-                    .return_type
-                    .map_or(any_type_ann(), |val| val);
-
-                let params: Vec<TsFnParam> = fn_expr
-                    .function
-                    .params
-                    .into_iter()
-                    .filter_map(|param| self.pat_to_ts_fn_param(param.pat))
-                    .collect();
-
-                Some(Box::new(TsType::TsFnOrConstructorType(
-                    TsFnOrConstructorType::TsFnType(TsFnType {
-                        span: fn_expr.function.span,
-                        params,
-                        type_ann: return_type,
-                        type_params: fn_expr.function.type_params,
-                    }),
-                )))
-            }
-            Expr::Arrow(arrow_expr) => {
-                let return_type = arrow_expr.return_type.map_or(any_type_ann(), |val| val);
-
-                let params = arrow_expr
-                    .params
-                    .into_iter()
-                    .filter_map(|pat| self.pat_to_ts_fn_param(pat))
-                    .collect();
-
-                Some(Box::new(TsType::TsFnOrConstructorType(
-                    TsFnOrConstructorType::TsFnType(TsFnType {
-                        span: arrow_expr.span,
-                        params,
-                        type_ann: return_type,
-                        type_params: arrow_expr.type_params,
-                    }),
-                )))
-            }
-            // Since fast check requires explicit type annotations these
-            // can be dropped as they are not part of an export declaration
-            Expr::This(_)
-            | Expr::Unary(_)
-            | Expr::Update(_)
-            | Expr::Bin(_)
-            | Expr::Assign(_)
-            | Expr::Member(_)
-            | Expr::SuperProp(_)
-            | Expr::Cond(_)
-            | Expr::Call(_)
-            | Expr::New(_)
-            | Expr::Seq(_)
-            | Expr::Ident(_)
-            | Expr::Tpl(_)
-            | Expr::TaggedTpl(_)
-            | Expr::Class(_)
-            | Expr::Yield(_)
-            | Expr::MetaProp(_)
-            | Expr::Await(_)
-            | Expr::Paren(_)
-            | Expr::JSXMember(_)
-            | Expr::JSXNamespacedName(_)
-            | Expr::JSXEmpty(_)
-            | Expr::JSXElement(_)
-            | Expr::JSXFragment(_)
-            | Expr::TsTypeAssertion(_)
-            | Expr::TsNonNull(_)
-            | Expr::TsInstantiation(_)
-            | Expr::PrivateName(_)
-            | Expr::OptChain(_)
-            | Expr::Invalid(_) => None,
-        }
-    }
-
-    fn decl_to_type_decl(&mut self, decl: &mut Decl) -> Option<()> {
-        let is_declare = self.is_top_level;
-        match decl {
-            Decl::Class(class_decl) => {
-                self.class_body_to_type(&mut class_decl.class.body);
-                class_decl.declare = is_declare;
-                Some(())
-            }
-            Decl::Fn(fn_decl) => {
-                fn_decl.function.body = None;
-                fn_decl.declare = is_declare;
-                self.handle_func_params(&mut fn_decl.function.params);
-                Some(())
-            }
-            Decl::Var(var_decl) => {
-                var_decl.declare = is_declare;
-
-                for decl in &mut var_decl.decls {
-                    if let Pat::Ident(ident) = &mut decl.name {
-                        if ident.type_ann.is_some() {
-                            decl.init = None;
-                            continue;
-                        }
-
-                        let ts_type = decl
-                            .init
-                            .take()
-                            .and_then(|init| self.expr_to_ts_type(init, false, true))
-                            .map(type_ann)
-                            .or_else(|| {
-                                self.mark_diagnostic_any_fallback(ident.span());
-                                Some(any_type_ann())
-                            });
-                        ident.type_ann = ts_type;
-                    } else {
-                        self.mark_diagnostic_unable_to_infer(decl.span());
-                    }
-
-                    decl.init = None;
-                }
-
-                Some(())
-            }
-            Decl::TsEnum(ts_enum) => {
-                ts_enum.declare = is_declare;
-
-                for member in &mut ts_enum.members {
-                    if let Some(init) = &member.init {
-                        // Support for expressions is limited in enums,
-                        // see https://www.typescriptlang.org/docs/handbook/enums.html
-                        member.init = if self.valid_enum_init_expr(init) {
-                            Some(init.clone())
-                        } else {
-                            None
-                        };
-                    }
-                }
-
-                Some(())
-            }
-            Decl::TsModule(ts_module) => {
-                ts_module.declare = is_declare;
-
-                if let Some(body) = ts_module.body.take() {
-                    ts_module.body = Some(self.transform_ts_ns_body(body));
-
-                    Some(())
-                } else {
-                    Some(())
-                }
-            }
-            Decl::TsInterface(_) | Decl::TsTypeAlias(_) => Some(()),
-            Decl::Using(_) => {
-                self.mark_diagnostic(DtsIssue::UnsupportedUsing {
-                    range: self.source_range_to_range(decl.span()),
-                });
-                None
-            }
-        }
-    }
-
-    fn transform_ts_ns_body(&mut self, ns: TsNamespaceBody) -> TsNamespaceBody {
-        let original_is_top_level = self.is_top_level;
-        self.is_top_level = false;
-        let body = match ns {
-            TsNamespaceBody::TsModuleBlock(mut ts_module_block) => {
-                self.transform_module_items(&mut ts_module_block.body);
-                TsNamespaceBody::TsModuleBlock(ts_module_block)
-            }
-            TsNamespaceBody::TsNamespaceDecl(ts_ns) => self.transform_ts_ns_body(*ts_ns.body),
-        };
-        self.is_top_level = original_is_top_level;
-        body
     }
 
     // Support for expressions is limited in enums,
@@ -723,252 +321,17 @@ impl FastDts {
         as_const: bool,
         as_readonly: bool,
     ) -> Option<Box<TsTypeAnn>> {
-        let span = expr.span();
-
-        if let Some(ts_type) = self.expr_to_ts_type(expr, as_const, as_readonly) {
+        if let Some(ts_type) = self.infer_type_from_expr(&expr, as_const, as_readonly) {
             Some(type_ann(ts_type))
         } else {
-            self.mark_diagnostic_any_fallback(span);
+            // self.mark_diagnostic_any_fallback(span);
             Some(any_type_ann())
         }
     }
 
-    fn class_body_to_type(&mut self, body: &mut Vec<ClassMember>) {
-        // Track if the previous member was an overload signature or not.
-        // When overloads are present the last item has the implementation
-        // body. For declaration files the implementation always needs to
-        // be dropped. Needs to be unique for each class because another
-        // class could be created inside a class method.
-        let mut prev_is_overload = false;
-
-        let new_body = body
-            .take()
-            .into_iter()
-            .filter(|member| match member {
-                ClassMember::Constructor(class_constructor) => {
-                    let is_overload =
-                        class_constructor.body.is_none() && !class_constructor.is_optional;
-                    if !prev_is_overload || is_overload {
-                        prev_is_overload = is_overload;
-                        true
-                    } else {
-                        prev_is_overload = false;
-                        false
-                    }
-                }
-                ClassMember::Method(method) => {
-                    let is_overload = method.function.body.is_none()
-                        && !(method.is_abstract || method.is_optional);
-                    if !prev_is_overload || is_overload {
-                        prev_is_overload = is_overload;
-                        true
-                    } else {
-                        prev_is_overload = false;
-                        false
-                    }
-                }
-                ClassMember::TsIndexSignature(_)
-                | ClassMember::ClassProp(_)
-                | ClassMember::PrivateProp(_)
-                | ClassMember::Empty(_)
-                | ClassMember::StaticBlock(_)
-                | ClassMember::AutoAccessor(_)
-                | ClassMember::PrivateMethod(_) => {
-                    prev_is_overload = false;
-                    true
-                }
-            })
-            .filter_map(|member| match member {
-                ClassMember::Constructor(mut class_constructor) => {
-                    class_constructor.body = None;
-                    self.handle_ts_param_props(&mut class_constructor.params);
-                    Some(ClassMember::Constructor(class_constructor))
-                }
-                ClassMember::Method(mut method) => {
-                    if let Some(new_prop_name) = valid_prop_name(&method.key) {
-                        method.key = new_prop_name;
-                    } else {
-                        return None;
-                    }
-
-                    method.function.body = None;
-                    if method.kind == MethodKind::Setter {
-                        method.function.return_type = None;
-                    }
-                    self.handle_func_params(&mut method.function.params);
-                    Some(ClassMember::Method(method))
-                }
-                ClassMember::ClassProp(mut prop) => {
-                    if let Some(new_prop_name) = valid_prop_name(&prop.key) {
-                        prop.key = new_prop_name;
-                    } else {
-                        return None;
-                    }
-
-                    if prop.type_ann.is_none() {
-                        if let Some(value) = prop.value {
-                            prop.type_ann = self
-                                .expr_to_ts_type(value, false, false)
-                                .map(type_ann)
-                                .or_else(|| Some(any_type_ann()));
-                        }
-                    }
-                    prop.value = None;
-                    prop.definite = false;
-                    prop.declare = false;
-
-                    Some(ClassMember::ClassProp(prop))
-                }
-                ClassMember::TsIndexSignature(index_sig) => {
-                    Some(ClassMember::TsIndexSignature(index_sig))
-                }
-
-                // These can be removed as they are not relevant for types
-                ClassMember::PrivateMethod(_)
-                | ClassMember::PrivateProp(_)
-                | ClassMember::Empty(_)
-                | ClassMember::StaticBlock(_)
-                | ClassMember::AutoAccessor(_) => None,
-            })
-            .collect();
-
-        *body = new_body;
-    }
-
-    fn handle_ts_param_props(&mut self, param_props: &mut Vec<ParamOrTsParamProp>) {
-        for param in param_props {
-            match param {
-                ParamOrTsParamProp::TsParamProp(param) => {
-                    match &mut param.param {
-                        TsParamPropParam::Ident(ident) => {
-                            self.handle_func_param_ident(ident);
-                        }
-                        TsParamPropParam::Assign(assign) => {
-                            if let Some(new_pat) = self.handle_func_param_assign(assign) {
-                                match new_pat {
-                                    Pat::Ident(new_ident) => {
-                                        param.param = TsParamPropParam::Ident(new_ident)
-                                    }
-                                    Pat::Assign(new_assign) => {
-                                        param.param = TsParamPropParam::Assign(new_assign)
-                                    }
-                                    Pat::Rest(_)
-                                    | Pat::Object(_)
-                                    | Pat::Array(_)
-                                    | Pat::Invalid(_)
-                                    | Pat::Expr(_) => {
-                                        // should never happen for parameter properties
-                                        unreachable!();
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                ParamOrTsParamProp::Param(param) => self.handle_func_param(param),
-            }
-        }
-    }
-
-    fn handle_func_params(&mut self, params: &mut Vec<Param>) {
-        for param in params {
-            self.handle_func_param(param);
-        }
-    }
-
-    fn handle_func_param(&mut self, param: &mut Param) {
-        match &mut param.pat {
-            Pat::Ident(ident) => {
-                self.handle_func_param_ident(ident);
-            }
-            Pat::Assign(assign_pat) => {
-                if let Some(new_pat) = self.handle_func_param_assign(assign_pat) {
-                    param.pat = new_pat;
-                }
-            }
-            Pat::Array(_) | Pat::Rest(_) | Pat::Object(_) | Pat::Invalid(_) | Pat::Expr(_) => {}
-        }
-    }
-
-    fn handle_func_param_ident(&mut self, ident: &mut BindingIdent) {
-        if ident.type_ann.is_none() {
-            self.mark_diagnostic_any_fallback(ident.span());
-            ident.type_ann = Some(any_type_ann());
-        }
-    }
-
-    fn handle_func_param_assign(&mut self, assign_pat: &mut AssignPat) -> Option<Pat> {
-        match &mut *assign_pat.left {
-            Pat::Ident(ident) => {
-                if ident.type_ann.is_none() {
-                    ident.type_ann =
-                        self.infer_expr_fallback_any(assign_pat.right.take(), false, false);
-                }
-
-                ident.optional = true;
-                Some(Pat::Ident(ident.clone()))
-            }
-            Pat::Array(arr_pat) => {
-                if arr_pat.type_ann.is_none() {
-                    arr_pat.type_ann =
-                        self.infer_expr_fallback_any(assign_pat.right.take(), false, false);
-                }
-
-                arr_pat.optional = true;
-                Some(Pat::Array(arr_pat.clone()))
-            }
-            Pat::Object(obj_pat) => {
-                if obj_pat.type_ann.is_none() {
-                    obj_pat.type_ann =
-                        self.infer_expr_fallback_any(assign_pat.right.take(), false, false);
-                }
-
-                obj_pat.optional = true;
-                Some(Pat::Object(obj_pat.clone()))
-            }
-            Pat::Rest(_) | Pat::Assign(_) | Pat::Expr(_) | Pat::Invalid(_) => None,
-        }
-    }
-
-    fn pat_to_ts_fn_param(&mut self, pat: Pat) -> Option<TsFnParam> {
-        match pat {
-            Pat::Ident(binding_id) => Some(TsFnParam::Ident(binding_id)),
-            Pat::Array(arr_pat) => Some(TsFnParam::Array(arr_pat)),
-            Pat::Rest(rest_pat) => Some(TsFnParam::Rest(rest_pat)),
-            Pat::Object(obj) => Some(TsFnParam::Object(obj)),
-            Pat::Assign(assign_pat) => {
-                self.expr_to_ts_type(assign_pat.right, false, false)
-                    .map(|param| {
-                        let name = if let Pat::Ident(ident) = *assign_pat.left {
-                            ident.sym.clone()
-                        } else {
-                            self.gen_unique_name()
-                        };
-
-                        TsFnParam::Ident(BindingIdent {
-                            id: Ident {
-                                span: assign_pat.span,
-                                ctxt: Default::default(),
-                                sym: name,
-                                optional: false,
-                            },
-                            type_ann: Some(type_ann(param)),
-                        })
-                    })
-            }
-            Pat::Expr(expr) => {
-                self.mark_diagnostic_unable_to_infer(expr.span());
-                None
-            }
-            // Invalid code is invalid, not sure why SWC doesn't throw
-            // a parse error here.
-            Pat::Invalid(_) => None,
-        }
-    }
-
-    fn gen_unique_name(&mut self) -> Atom {
+    fn gen_unique_name(&mut self, name: &str) -> Atom {
         self.id_counter += 1;
-        format!("_dts_{}", self.id_counter).into()
+        format!("{name}_{}", self.id_counter).into()
     }
 }
